+++ players/player.php
@@ -1,9 +1,9 @@
 <?php
 $response = array();
 
-require(dirname(__FILE__) . "/../classes/class.Player.inc.php");
+include(dirname(__FILE__) . "/../classes/class.Player.inc.php");
 
-$player = new \WurmUnlimitedAdmin\PLAYER();
+$player = new PLAYER();
 
 $playerList = $player->GetPlayers($_POST);