+++ servers/view/view.php
@@ -3,7 +3,7 @@ $response = array();
 
 require(dirname(__FILE__) . "/../../classes/class.Server.inc.php");
 
-$server = new \WurmUnlimitedAdmin\SERVER();
+$server = new SERVER();
 
 $serverList = $server->GetServers($_POST["serverID"]);